+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.44.0] - 2021-04-12
 
+### Added
+- DropDown: Added class disabled
+
 ### Fixed
 - NumericTextBox: Fixed rounding numbers on change event